Update code to support StateMachine (wip)
Currently just a little refactor of the Voxaudio.hpp but the goal here is to have a state machine per sound (or more like per channel or something like that). So I need a bit more data and other things. So, it's slowly coming together.
Next step will be the Fmod::Studio API so the sound designers can use FmodStudio and make the whole event this way.

I will probably need a way to set more data. Like a KeyPair thing but for now it's enought.